mds: send unlinked dentry's inode attributes in the reply trace for unlink
Currently when a dentry is unlinked, the reply doesn't contain any
target inode information. That's usually fine as it's unreachable
in many cases anyway, but if the inode is still hardlinked to a
different dentry, then it is still reachable.

Worse, the client could hold a lease on the remaining dentry. If the
application calls ceph_ll_lookup, libcephfs will satisfy the lookup out
of its cache, and fill out the stat structure with (now outdated)
information from its cached inode.

This is a partial fix for this problem, which makes the MDS send the
target inode information in the trace after an unlink. This should be
harmless -- in the event that we discard the inode the new attrs will
just get discarded as well.

Signed-off-by: Jeff Layton <jlayton@redhat.com>